Check project privacy status
As data hubs should be updated in line with the umbrella project, additional
information is now outputted by this script. A new column in the TSV file indicates
the status of the project and a new file containing only public project accessions
can be used to import public projects to the datahub.